Two new attributes generated
In the XES file, two new fileds have been added: RelationShipFrom and RelationShipTo, which respectively indicate the name of the class from which the relationship originates and the class to which the relationship is directed. Additionally, the fields RelationShipFromId and RelationShipToId, which represent the ID of the class from which the relationship originates and the ID of the class to which the relationship is directed, are also present.