Class diagrams generation
Added the functionality to generate class diagram parsing a JSON File.
Finally managed the handling of auxiliary and master views. Now, for each class, only attributes, operations, and relations in the current view are described. Additionally, included information on the project and added a window to inform the user about the export outcome.